+++ packages/sdk/src/index.ts
@@ -12,20 +12,8 @@ export type {
 	CustomizeMutation,
 	CustomizeSubscription,
 } from './configure/operations';
-export { configureWunderGraphServer } from './server';
 export type { GraphQLOperation } from './graphql/operations';
 export type { Template, TemplateOutputFile } from './codegen/index';
-export type {
-	WebhookHttpEvent,
-	WebhookHttpResponse,
-	Webhook,
-	WebhookRequestContext,
-	WebhookLogger,
-	WebhookHeaders,
-	WebhookQuery,
-	WebhookConfiguration,
-} from './webhooks/types';
-export type { WebhookVerifierConfiguration } from './webhooks/verifiers';
 export { default as templates } from './codegen/templates';
 export { introspect, createMockApi, Api } from './definition';
 export { configureWunderGraphApplication } from './configure';
@@ -42,6 +30,5 @@ export {
 	TypeScriptResponseModels,
 } from './codegen/templates/typescript/index';
 export { TypeScriptClient } from './codegen/templates/typescript/client';
-export { GithubWebhookVerifier, CreateWebhookVerifier, WebhookVerifierKind } from './webhooks/verifiers';
 export { PlaceHolder } from './configure/variables';
 export { EnvironmentVariable } from './configure/variables';
