+++ packages/sdk/src/server/index.ts
@@ -1,6 +1,4 @@
 export type { HooksConfiguration } from './types';
-
-export { configureWunderGraphServer } from './server';
 export type { OperationArgsWithInput, InternalClient } from './internal-client';
 export type { GraphQLServerConfig } from './plugins/graphql';
 
@@ -39,4 +37,22 @@ export type {
 	FetchConfig,
 } from './client';
 
+export type {
+	WebhookHttpEvent,
+	WebhookHttpResponse,
+	Webhook,
+	WebhookRequestContext,
+	WebhookLogger,
+	WebhookHeaders,
+	WebhookQuery,
+	WebhookConfiguration,
+} from '../webhooks/types';
+export type { WebhookVerifierConfiguration } from '../webhooks/verifiers';
+
 export type { FastifyLoggerInstance } from 'fastify';
+
+export { configureWunderGraphServer } from './server';
+
+export { GithubWebhookVerifier, CreateWebhookVerifier, WebhookVerifierKind } from '../webhooks/verifiers';
+
+export { EnvironmentVariable } from '../configure/variables';
